Feat/multiple own mons 259

+++ src/main/java/de/uniks/beastopia/teaml/rest/AbilityDto.java
@@ -6,7 +6,7 @@ public record AbilityDto(
         String description,
         String type,
         int maxUses,
-        int accuracy,
+        double accuracy,
         int power
 ) {
 }
